Added alias logic
