+++ src/main/java/com/example/shorturl/controller/api/ShortenUrlControllerV1.java
@@ -1,8 +1,9 @@
 package com.example.shorturl.controller.api;
 
-import com.example.shorturl.model.url.UrlEntity;
+
 import com.example.shorturl.model.payload.request.url.UrlRequest;
 import com.example.shorturl.model.payload.response.url.UrlResponse;
+import com.example.shorturl.model.url.UrlEntity;
 import com.example.shorturl.service.url.ShortenUrlService;
 import com.example.shorturl.utils.exceptions.UrlNotFoundException;
 import org.apache.commons.lang3.StringUtils;
@@ -18,8 +19,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -49,30 +48,6 @@ public class ShortenUrlControllerV1 {
                         .originalUrl(url.getOriginalUrl())
                         .expirationDate(url.getExpirationDate())
                         .build(), HttpStatus.OK);
-    }
-
-    @GetMapping("/redirect/{shortUrl}")
-    public ResponseEntity<UrlResponse> redirectUrl(@PathVariable String shortUrl, HttpServletResponse response) throws UrlNotFoundException, IOException {
-
-        HttpHeaders headers = new HttpHeaders();
-
-        if(StringUtils.isEmpty(shortUrl))
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Url request is missing or empty");
-
-        Optional<UrlEntity> optionalUrl = shortenUrlService.getEncodedUrl(shortUrl);
-
-        UrlEntity url = optionalUrl.orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Url not found"));
-
-        if(url.getIsDeleted())
-            throw new ResponseStatusException(HttpStatus.GONE, "Url has been deleted");
-
-        if(url.getIsExpired())
-            throw new ResponseStatusException(HttpStatus.GONE, "Url has expired");
-
-        //TODO: Handle IOException
-        response.sendRedirect(url.getOriginalUrl());
-
-        return ResponseEntity.status(HttpStatus.OK).headers(headers).build();
     }
 
     @GetMapping("/{id}")
